% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateThreshold.R
\name{evaluateThreshold}
\alias{evaluateThreshold}
\title{Evaluate Thresholds for Network Clustering}
\usage{
evaluateThreshold(input, seed = 42)
}
\arguments{
\item{input}{A data frame containing terms, genes, and additional metadata.}
}
\value{
A Plotly object visualizing the evaluation of thresholds.
}
\description{
Evaluates graph properties (connected components, modularity, cluster count) for a range of thresholds.
}
