% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{plotBubbleplot}
\alias{plotBubbleplot}
\title{Plot Enrichment Bubble plot}
\usage{
plotBubbleplot(
  input,
  plot_colors = list(up = c("#ececec", "#e17ecd", "#7900d5"), down = c("#ececec",
    "#41B7C4", "#2A5783"), default = c("#ececec", "#41B7C4", "#2A5783"))
)
}
\arguments{
\item{input}{A data frame containing enrichment results with columns \code{Cluster_Annotation}, \code{condition}, \code{pval_pooled}, and optionally \code{terms_per_cluster}.}

\item{plot_colors}{List. A named list of color palettes for the heatmap. Should contain \code{up}, \code{down}, and \code{default}.}
}
\value{
A ggplot object
}
\description{
Generates a Bubble plot of enrichment results, with options to split by regulation
}
