% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizer.R
\name{summarizer}
\alias{summarizer}
\title{Show summary statistic in a widget}
\usage{
summarizer(
  data,
  statistic = c("count", "sum", "mean"),
  column = NULL,
  statistic2 = c("count", "sum", "mean", NULL),
  column2 = NULL,
  selection = NULL,
  digits = 0,
  prefix = NULL,
  suffix = NULL,
  width = 250,
  height = 65,
  elementId = NULL
)
}
\arguments{
\item{data}{Data to summarize, normally an instance of \link[crosstalk:SharedData]{crosstalk::SharedData}.}

\item{statistic}{The statistic to compute.}

\item{column}{For \code{sum} and \code{mean} statistics, the column of \code{data} to summarize.
Not used for \code{count} statistic.}

\item{statistic2}{Optional. If provided, summarises parameter \code{column2}, which becomes
a denominator of \code{column}.}

\item{column2}{Optional. For \code{sum} and \code{mean} statistics, the column of \code{data} to summarize.
Not used for \code{count} statistic.}

\item{selection}{Expression to select a fixed subset of \code{data}. May be
a logical vector or a one-sided formula that evaluates to a logical vector.
If used, the \code{key} given to \link[crosstalk:SharedData]{crosstalk::SharedData} must be a fixed column (not row numbers).}

\item{digits}{Number of decimal places to display, or NULL to display full precision.}

\item{prefix}{text to be placed before the result number (e.g '$').}

\item{suffix}{text to be placed after the result number (e.g '\%', '/month').}

\item{width}{width of the widget (in px).}

\item{height}{height of the widget (in px).}

\item{elementId}{custom HTML ID for the returning widget.}
}
\description{
A \code{summarizer} displays statistic derived from a linked table.
Its primary use is with the \code{crosstalk} package. Used with \code{crosstalk},
a \code{summarizer} displays a value which updates as the data selection
changes. It summarizes one-dimentional as well as two-dimensional data when \code{column2}
and \code{statistic2} is provided
}
