% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRUplotIgraph.R
\name{TRUplotIgraph}
\alias{TRUplotIgraph}
\title{Generate an Interactive Network Graph with Repelled Labels}
\usage{
TRUplotIgraph(input, ts, seed = 42)
}
\arguments{
\item{input}{A data frame containing terms, genes, and additional metadata.}

\item{ts}{A numeric threshold for edge inclusion based on distance weights.}
}
\value{
A Plotly object representing the network graph.
}
\description{
Creates a Plotly-based interactive graph visualization of terms and their relationships, minimizing label overlap.
}
