% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxSummarizer.R
\name{valueBoxSummarizer}
\alias{valueBoxSummarizer}
\title{Crates a Flexdashboard-like valuebox for object returned by \code{\link[=summarizer]{summarizer()}} function}
\usage{
valueBoxSummarizer(
  value,
  caption = NULL,
  icon = NULL,
  color = NULL,
  href = NULL
)
}
\arguments{
\item{value}{string, numeric value or span object returned by \code{\link[=summarizer]{summarizer()}}.}

\item{caption}{text placed on the bottom of the valuebox.}

\item{icon}{class of a font-awesome icon. Visit: fontawesome.com for more info.}

\item{color}{background color of the valuebox.}

\item{href}{hyperlink string.}
}
\description{
A \code{summarizer} displays statistic derived from a linked table.
Its primary use is with the \code{crosstalk} package. Used with \code{crosstalk},
a \code{summarizer} displays a value which updates as the data selection
changes. It summarizes one-dimentional as well as two-dimensional data when \code{column2}
and \code{statistic2} is provided
}
