% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetsummarize.R
\name{widgetsummarize}
\alias{widgetsummarize}
\title{Show a single summary statistic in a widget}
\usage{
widgetsummarize(
  data,
  statistic = c("count", "sum", "mean"),
  column = NULL,
  statistic2 = c("count", "sum", "mean", NULL),
  column2 = NULL,
  selection = NULL,
  digits = 0,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Data to summarize, normally an instance of \link[crosstalk:SharedData]{crosstalk::SharedData}.}

\item{statistic}{The statistic to compute.}

\item{column}{For \code{sum} and \code{mean} statistics, the column of \code{data} to summarize.
Not used for \code{count} statistic.}

\item{statistic2}{Optional. If provided, summarises parameter \code{column2}, which becomes
a denominator of \code{column}.}

\item{column2}{For \code{sum} and \code{mean} statistics, the column of \code{data} to summarize.
Not used for \code{count} statistic.}

\item{selection}{Expression to select a fixed subset of \code{data}. May be
a logical vector or a one-sided formula that evaluates to a logical vector.
If used, the \code{key} given to \link[crosstalk:SharedData]{crosstalk::SharedData} must be a fixed column (not row numbers).}

\item{digits}{Number of decimal places to display, or NULL to display full precision.}
}
\description{
A \code{widgetsummarize} displays a single statistic derived from a linked table.
Its primary use is with the \code{crosstalk} package. Used with \code{crosstalk},
a \code{widgetsummarize} displays a value which updates as the data selection
changes. For percentages and weighted average summaries, provide parameters 'statistic2'
and 'column2'.
}
\seealso{
\url{https://kent37.github.io/summarywidget}
}
